backend/requirements.txt
Django==3.2.17
django-filter==21.1
djangorestframework==3.14.0
djoser==2.1.0
drf-extra-fields==3.4.1
gunicorn==20.1.0
Pillow==9.4.0
reportlab==3.6.12
psycopg2-binary==2.9.5